AIservices/zega/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
chromadb==0.5.23
langchain-core==0.3.28
langchain-google-genai==2.0.8
google-generativeai==0.8.3
